+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changes
 - Rename `async_upnp_client.traffic` logger to `async_upnp_client.traffic.upnp` and add `async_upnp_client.traffic.ssdp` logger
 - Added `DeviceUpdater` to support updating the `UpnpDevice` inline on changes to `BOOTID.UPNP.ORG`/`CONFIGID.UPNP.ORG`/`LOCATION`
 - Added support for PAUSED_PLAYBACK state (#56, @brgerig)
+- Add `DmrDevice.transport_state`, deprecate `DmrDevice.state`
 
 
 0.15.0 (2021-03-13)